Change default of `in_specs` argument of `shard_map` from `None` to `Infer`.
**Why?**

*Chapter 1: The story of defaults*

`in_specs=None` in `shard_map` meant 2 things:

* If in Auto mode, it meant the args were static
* If in Explicit mode, it meant to infer the in_specs from arguments

This lead to inconsistent and confusing behavior. So what do we do?

Thankfully, shard_map's younger brother, `smap` already has the solution for us!

*Chapter 2: smap to the rescue*

`in_axes` argument of `smap` defaults to `Infer` which means it will try to infer the arguments and error if it can't.

So `smap` uses `Infer` to indicate "infer from arguments" and `None` to indicate "static argument".

*Chapter 3: Copy paste*

We can use the `smap` semantics and apply them to `shard_map` to disambiguate "infer from args" and "static args"

Hence this change changes the default of `in_specs` from `None` to `Infer`. So in `shard_map`:
* in_specs=None: means argument is static
* in_specs=Infer (default): means infer from argument if context mesh is Explicit.

This also matches the behavior of `smap` making `shard_map` and `smap` consistent.

PiperOrigin-RevId: 791003495