[JAX] Define `PyUserContext`
`jax::PyUserContext` is a subclass of `xla::ifrt::UserContext`. When it is set
up as an active user context by using `xla::ifrt::UserContextScope`, it will be
associated with a newly created `xla::ifrt::Array` (which `jax::PyArray` wraps)
and `xla::ifrt::LoadedExecutable` (which `jax::PyLoadedExecutable` wraps). This
allows IFT implementations to provide the contextual information in error
formatting or live runtime object queries.

This change only defines `jax::PyUserContext` in jaxlib. A subsequent change
will use it instead of `jax::Traceback::Get()` when constructing `jax::PyArray`
and `jax::PyLoadedExecutable`.

PiperOrigin-RevId: 791848502