Remove NumPy version pin in CI.
As far as I can tell this pin existed to work around a crash in CI for Linux aarch64 builds due to a numpy bug. The bug appears fixed in 2.3, and the pin was accidentally being applied for 3.14-nogil causing test failures in that build. We should be able to just remove it.