[Mosaic] Refine scalar load test for narrow bitwidths
This CL modifies the test "test_narrow_bitwidth_scalar_prefetch." The test was originally introduced to isolate and reproduce a 1D packed memref load issue. However, the existing implementation also performs a vector addition to data in VMEM, which triggers an unrelated issue, obscuring the test's primary focus.

Address this by modifying the test to concentrate solely on the 1D load and broadcast functionality by removing the vector addition component.

Additionally, add `int16` to the set of supported dtypes for this test.

PiperOrigin-RevId: 794074522