Add internal helper functions for computing default dtypes.
Use default dtype helpers in place of canonicalize_dtype(int), etc.

Enable-x64 mode current has several roles, one of which is choosing the default dtypes for a number of jax APIs. By moving the default dtype choice into its own internal API, we are free to experiment with separating these.

I did not merge these with the default_dtype_bits experiment, since (a) we want to defer the choice to after module import and (b) those are part of a dead experiment.

PiperOrigin-RevId: 794119254