[jaxpr consts] Fix AOT execution with 64-bit captured constants
Add a test to check the behavior when the code is AOT lowered with
a different value of the `jax_enable_x64` configuration value
than it is executed for.

Previously, when the constants are embedded as `stablehlo.constant`
we can use diffenent values of the `jax_enable_x64` flag. To
keep this behavior we must take some care when we invoke the
executable in AOT mode.