[JAX] Pre-compute hash for `jax::CallSignature` at construction time for C++ exception safety
`jax::CallSignature` is used as a key of a C++ unordered container and its hash
would be used. Previously, it invoked an associated `AbslHashValue`, which can
throw a C++ exception if some member of `jax::CallSignature` is not hashable.
The problem is that the calling context of hash computation may not necessarily
be C++ exception safe (e.g., ABSL containers).

This change avoids this unsafe hash function problem by pre-computing the hash
value when the key `jax::CallSignature` is constructed. Accessing the cached
hash value will never throw.

As a good side effect, the hash value of `jax::CallSignature` will be computed
only once. This will make rehashing the container content faster.

A downside will be that the hash value must be manually filled in during
`jax.CallSignature` construction.

PiperOrigin-RevId: 794760351