Attempt to fix flaky test
We are seeing flakyness in lax_control_flow_test::test_cond_memory_leak.
The failures are in the conditional `assertEqual(base, nbufs())`
of the form `7 != 2` and `8 != 2`. It seems that the value
`base` includes too many live buffers. My best guess is that since
this started when I added `gc.collect()` inside `leak()`, this may
be fixable if we also call `gc.collect()` just before computing
`base`.