[Mosaic GPU] Add a `Reshape` expression constructor
This will be helpful to propagate layouts through `shape_cast`. Reducing the
reshape of a layout works as follows:

* Given a `WGStrided` layout, simply change the underlying shape if the
  shapes are compatible (have the same number of elements). Otherwise, the
  reduction is unsatisfiable;
* Given a `WGSplat` layout, do the same as in the `WGStrided` case;
* Given a `TiledLayout`, propagate the layout if the new shape is still
  tilable by the layout in a way that preserves the mapping of the data to
  lanes and warps---which is guaranteed if all the tiled dimensions (except
  for the majormost ones) are unmodified, and the majormost dimension remains
  tilable. If this is not possible, then the reduction is undetermined.

In the `TiledLayout` case, it is possible that we could perform a reshape of
the base tile shape and subsequent tilings so as to make the reshape work out.
We don't implement that logic now, but this is the reason for not making that
case unsatisfiable.

PiperOrigin-RevId: 795032848