Use Python 3.14 context aware warnings if they are enabled.
We don't need our own bespoke warning infrastructure now that https://docs.python.org/3.14/whatsnew/3.14.html#concurrent-safe-warnings-control has landed.

Sadly it's only enabled in free threaded builds, but it's a problem even in non-free-threaded builds. But since the behavior is controlled by an interpreter flag there's no way for us to enable it from within the program if it's not already enabled.