[Mosaic TPU] Canonicalize rank-changing broadcasts to reshape + broadcast
This allows us to simplify (or rather avoid complicating) the layout rules
for broadcasts. The layout rules for reshape already deal with e.g. removing
implicit dims when possible when new singleton dimensions are added and we
can take advantage of that to infer better layouts.

This used to happen automataically, because the Pallas:TPU lowering pretty much
always inserts a shape_cast before a broadcast, but the recently added MLIR
folding rules started removing shape_casts that can be "fused" into the broadcast.

PiperOrigin-RevId: 795398502