[Mosaic] Fix deadlock in 1D packed memrefs test
The test `test_narrow_bitwidth_scalar_prefetch` is currently decorated with `@only_passes_in_interpret`, which requires the test to throw an exception to pass.

This creates a deadlock. Implementing the required 1D packed memref logic prevents the exception, which in turn causes the decorated test to fail. However, we cannot remove the decorator first, as the test would then fail due to the unhandled exception.

Because the two necessary changes (in `lower_to_llo.cc` and `tpu_pallas_test.py`) cannot be submitted atomically, they are mutually blocked.

Resolve the deadlock by removing the decorator and instead use the `is_cloud_tpu_older_than()` function to conditionally run the test.

PiperOrigin-RevId: 796859066