[Pallas:MGPU] Only synchronize the SMEM->GMEM copies in pipeline if there was something to copy
We always use plgpu.pipeline as an implementation detail in pallas_call, even if all the
outputs have a GMEM (or ANY) memory space. In those situations it's pointless to emit the
epilogue that awaits the copies. It's not a huge deal, but it is a few barriers and other ops.

PiperOrigin-RevId: 797227909