[Pallas:MGPU] Disallow outputs in WG specialized pipelines with multiple compute warpgroups
The current implementation doesn't ensure proper synchronization between the compute
warpgroups and the writebacks to HBM. It is, however, safe if there is only a single
compute warpgroup. As far as I can tell we don't actually have any real kernels that
would use this functionality so I'm just locking it down for now instead of writing
a full fix, which would be much more complicated.

PiperOrigin-RevId: 797258010