Use a fixed set of monoids.
This commit changes from allowing users to provide custom monoids to
using a fixed set of predefined monoids. This simplifies a couple
things. For example, it makes it clearer how we hash and cache things.
Users also don't have to provide an identity element anymore. In the
future, we can generalize the API if needed.