Canonicalize `dtype` arg to `_get_and_check_dtype`.
This fixes a bug in `make_array_from_callback` where the shard data dtype was canonicalized but the `dtype` arg was not, leading to a potential error if the `dtype` arg was passed as the (NumPy) dtype of the input shards. Rearranged logic in `_get_and_check_dtype` to avoid calling `canonicalize_dtype` twice.

PiperOrigin-RevId: 797529921