#sdy remove the requirement that all the in/out shardings of the initial call to JAX export - the part that saves the model into a serialized `Exported` object.
Note that there was no such enforcement on the loaded exported function. You could have always saved a program all with meshes A and load it with meshes B. Technically even load+jax.jit it with meshes A/B/C/D/etc since `jax.jit` doesn't enforce the same mesh. This CL just relaxes the call to save a program, allowing for any set of meshes, matching the contract of `jax.jit`.

PiperOrigin-RevId: 797772968