Add support for building JAX CUDA artifacts with multiple CUDA versions
This change introduces the ability to build JAX CUDA artifacts against different major versions of CUDA by parameterizing the CI build scripts. This is a prerequisite for adding CUDA 13 support. A new environment variable `JAXCI_ARTIFACT_CUDA_MAJOR_VERSION` is introduced, defaulting to 12.

The `build_artifacts.sh` script is updated to use this variable to construct the correct Bazel configuration (e.g., ..._cuda12, ..._cuda13) and also pass the version to the build CLI.

PiperOrigin-RevId: 798246764