# [PST:GA] Add `key_value_increment` to `sync.py` and update async checkpointing to use the kv-store `execution_id`
This change just updates `checkpoint.py` to use the `execution_id` stored in the kv-store rather than the chief provided one.

It should be a no-op, as async checkpointing will still be chief-only. In a subsequent CL the restriction on async checkpointing only being triggerable on the PST chief can be removed.

PiperOrigin-RevId: 798366673