[Mosaic GPU] Unify the logic to conjure new assignments from hints or default variables in layout inference.
This allows having a single-stop shop to extract new assignments, and removes
duplicated logic.

We no longer remove hints explicitly when introducing a new assignment. That's
OK, the call to `reduce_hints` in the recursive call should take care of
cleaning that up anyway.

I had to update a test because I changed the traversal order (previously, hints
were traversed in reverse order). The test had a comment documenting its
reliance on traversal order, so that is not unexpected.

PiperOrigin-RevId: 799642124