[Mosaic:TPU][Relayouts] Always use replicated/gather retiling for 32-bit (1, 128) -> (n, 128) retiling with replicated 2nd minor
When the shape's 2nd minor is 1, for (1, 128) tiling, a replicated 2nd minor is the same as a non-replicated one. However, previously, we would only hit into this relayout when the source was _not_ marked as replicated on earlier gens.

The logic was strange and I believe was only there because of missing handling of replication for implicit dim changes.

PiperOrigin-RevId: 799729767