Drop support for NumPy 1.26 and SciPy 1.12.
Per our deprecation policy, support for NumPy 1.26 can be dropped in September 2025, i.e., in our next release.

Normally SciPy 1.12 would be supported longer. But SciPy 1.12 is incompatible with NumPy 2.0, so if we're bumping the minimum NumPy version, we might as well also bump the minimum scipy version to 1.13 which is the first NumPy 2.0 compatible version.

PiperOrigin-RevId: 800069859