[pallas:mgpu] warp specialized pipeline with special delay_release.
Instead of the compute warps arriving to the barriers with a delay, the memory
warp is fetching into slots with a delay. Since not all inputs have the same
delay release, if we were to keep the delays being handled on the compute side
we would need to maintain different barriers for each input.

Before this CL the consume barriers would be arrived to at every step, making it
necessary ti wait on them by the memory at the end. We remove this requirement
in the non-manual barriers case by only arriving to the barrier on the steps
that we need to do a fetch, ie

   step \in [min(delay_release), num_steps - stages + max(delay_release))

This is not possible in manual consume barriers.

PiperOrigin-RevId: 802143932