[Pallas:MGPU] Add support for dovetailing multiple WS pipelines
The previous implementation always caused a bubble, because the memory WG
always awaited the completion of all compute steps before continuing. Now,
with `pipeline_state` it is possible to compose multiple sequential invocations
of WS pipelines in a way that does not create any memory bubbles but still makes
it possible to quiesce all operations if desired.

PiperOrigin-RevId: 802525515