Make `roofline` BUILD target & add `call_tf`.
The primary goal of this change is to add `call_tf` as a "standard" (i.e. 0 cost) op. It counts for 0 since we cannot know the cost of the underlying TF function.

However, importing `jax.experimental.jax2tf.call_tf` to roofline is tricky given the current BUILD setup. Thus, we add build targets for `roofline` and add the needed dependencies directly.

This also helps us clarify which parts of the codebase do what, and which parts are public/private.

PiperOrigin-RevId: 802636454