[Mosaic TPU] Improve error message for unsupported AddI vector types
The `arith::addi` operation fails with a confusing message when adding integer vectors with element types other than 16 or 32 bits.
Improve that by introducing a new check in the canonicalization pass to verify the vector's element type. If the type is unsupported, it now emits a clear and helpful error message.

Additionally, a new test, `test_vector_integer_addition`, is added to verify the correct behavior.

PiperOrigin-RevId: 802701184