Change default of `enable_sharding_check` in `typematch` to True and remove that option.
The current sharding check only works if any axis is explicit in t1 or t2. TODO is to make it work with Auto and Manual mode too.

This check is already paying off since it caught a bug in the scatter sharding propagation where the `update` was not getting broadcasted to the correct sharding. Also scatter sharding rule errors out now if operand spec and update spec don't match up.

PiperOrigin-RevId: 804680103