[mosaic_gpu] Added a pass which "resolves" trivial locations in a kernel
JAX recently started caching MLIR lowerings, so a `pallas_call` is first lowered
without any incoming traceback information and then inlined at a concrete call
site with traceback information patched in.

This scheme does not work for the kernel module, which is serialized in the
`backend_config=` of a `custom_call` in the lowered representation of a
`pallas_call`.

The new pass finds all "pallas_call" locations and replaces them with a
concrete location derived from the nearby ops.

PiperOrigin-RevId: 805794486