[Pallas/Mosaic GPU] Fix the warp specialized pipeline to handle the case where `delay_release` is used and `max_concurrent_steps` is not a divisor of the number of steps.
In that case, we were not computing the slots of the barriers to await
correctly, which ended up causing deadlocks.

Since pipelines of pipelines are not supported when the number of steps is not
divisibile by `max_concurrent_steps`, the issues doesn't show up there---but we
note that the current logic would have to be changed if we ever wanted to
support that case.

PiperOrigin-RevId: 805806113