[JAX] Apply `xla::ifrt::ExpandUserContexts()` to the locations that can surface propagated errors through asynchronous operations
Runtimes can propagate errors through asynchrously dispatched operations
(executions, transfers) and surface the errors when the user requests the array
content (through `xla::ifrt::Array::CopyToHostBuffer()` or requests the array
status via a ready future (e.g., `xla::ifrt::Client::GetReadyFutures()`
`xla::ifrt::ExecuteResult::status`) synchronously. The runtimes would use
`xla::ifrt::AttachUserContextRef` to annotate `absl::Status` with a
`UserContext` that represents the context of the culprit operation.

This change adds `xla::ifrt::ExpandUserContexts()` to the known locations in
JAX where the errors from async operations can be surfaced synchronously. Any
IFRT `UserContext` attached to an `absl::Status would be expanded and appended
to the end of status message so that normal `absl::Status` handler can show the
error to the user.

If `absl::Status` is an OK or has no `UserContext` attached, this change is
no-op; the same status will be returned without any expansion.

For the background on the reason why `xla::ifrt::ExpandUserContexts()` is
called on the IFRT user (JAX) side: the user can control when it is expanded
and can pick the best time when it can be done efficiently (e.g., expanding
`jax::PyUserContext` would require GIL as it dumps the associated Python
traceback, so having the expansion done while GIL is already being held is the
best time.

PiperOrigin-RevId: 806309985