[Mosaic] Verify scalar IOps types
We only support scalar integer operations on i32. Instead of adding a check for every instruction, introduce a single check in the canonicalization pass (in `canonicalize_mosaic.cc`).
Note: Currently, only deal with arith.addi; for readability, remaining operations will be added in a follow-up CL.

To verify the behavior for all scalar integer operations (of which there are many), we are grouping all related tests into a new class: PallasScalarIOpsTest.

For now, move the existing `test_scalar_integer_addition` function (which checks `arith.addi`) into this new class and rename it. The subsequent CL will add a dozen more tests to this class.

PiperOrigin-RevId: 807834144