[Mosaic] Verify more scalar IOps
Previously, a check was added to the canonicalization pass to verify
that scalar integer ops use the `i32` type (the only one supported
for those ops). However, this check was only applied to `arith.addi`.

Extends this i32 type verification to a wider set of scalar integer
operations and add new tests to validate the change.

PiperOrigin-RevId: 807849562