[Mosaic] Add a new kernel-wide compiler flag `shape_invariant_numerics` to pallas/mosaic. Default is false.
If set to be true, the following will hold:
1. operations applied to operands with a particular vector type will always produce exact the same result.
2. if an operation is a map along some dimension (i.e. it is equivalent to splitting the input along that dim, doing the operation on each slice independently and then concatenating) then its semantics don't depend on the size of the map dimension.

Currently, we will replace all-reduce across lanes and sublanes by multiplication when data is replicated and the reduction kind is sum. We use the newly introduced flag to guard this optimization. In future, more optimizations can be controlled by this flag.

PiperOrigin-RevId: 808807691