[Mosaic] Add a new kernel-wide compiler flag `consistent_numerics` to pallas/mosaic. Default is false, which means some optimizations will be applied while things like reduction order is not guaranteed. So, properties like batch-invariance don't generally hold.
If set to be true, more consistent numerics will be supported, e.g., reduction order will be more deterministic.

Currently, we will replace all-reduce across lanes and sublanes by multiplication when data is replicated and the reduction kind is sum. We use the newly introduced flag to guard this optimization. In future, more optimizations can be controlled by this flag.

PiperOrigin-RevId: 808807691