[Mosaic] Refactor: Unify builder usage in canonicalize_elementwise
The function `canonicalize_elementwise()` was the only user of `OpBuilder` in the file, deviating from the convention of using `CanonicalBuilder`. This was because `canonicalize_elementwise` needs to create operations dynamically using their name, which wasn't supported by the existing templated `create` methods in `CanonicalBuilder`.

To improve consistency and readability, introduce a new non-templated `create` method to the `CanonicalBuilder` class. This new method allows creating operations by passing the operation name as a string. The `canonicalize_elementwise()` function is updated to use this new method, aligning its implementation with the rest of the file.

The change requires the inclusion of "mlir/IR/BlockSupport.h" and "mlir/IR/ValueRange.h". Additionally, due to commit 2eb733b ("[mlir][tblgen] add concrete create methods") in LLVM, which moved `ImplicitLocOpBuilder` from `mlir/IR/ImplicitLocOpBuilder.h` to `mlir/IR/Builders.h`, the unnecessary inclusion of `ImplicitLocOpBuilder.h` is removed.

PiperOrigin-RevId: 809257786