[Mosaic] Support small second minor tiling for 16 bit reshape.
This can be by mixed unpacking/packing to halve/double sublane tiling. Alternatively, we can unpack to 32 bit and to sublane shuffle, but we lack of support for general tpu::SublaneShuffleOp. And it's not going to be faster IMO as arbitrary shuffle pattern needs to be emulated by store + load.

I think this can be extended 8 bit, from (32, 128) -> (8, 128) -> (2, 128) tiling. But I haven't come up with an efficient pattern for other tiling.

PiperOrigin-RevId: 809273967