[Mosaic GPU] Saturate tiling constraints for equal vars.
We need this in order to be able to conjure up the correct tilings for `subview` ops. With `subview` we will have two different vars for the source and the result, but both have the same transforms. There are cases (e.g. ragged dot) where the constraints are gathered on the one variable, but we need to conjure up an assignment for the other.

PiperOrigin-RevId: 810891732