[Pallas/interpreter] Correctly use `num_cores` (instead of `num_devices`) when generating a random virtual device id.
The previous use of `num_devices` used to mask a crash of a test case. With this change, the same test case would crash now. To avoid this, we make `vector_clock_size` an additional interpreter parameter that defaults to twice the total number of cores.

In accordance with this change, the global (constant) variable `NUM_VIRTUAL_CORES` is renamed to `_vector_clock_size` and is set based on the added interpreter parameter.

PiperOrigin-RevId: 811470482