[Pallas:MGPU] Infer `optimized` attribute for Pallas `swap` lowering.
We do not set the `optimized` attribute for `vector.load` and `vector.store` instructions in the Pallas `swap` lowering. When lowering `vector.load` and `vector.store` in MGPU, if the `optimized` attribute is not set we first try emitting an optimized transfer and defaults to a non-optimized transfer on failure.

PiperOrigin-RevId: 811780848