[JAX] Add a serialization roundtrip test for colocated Python
This change adds a test for serialization roundtrip that the implementation of the colocated Python API rely.

As part of this change, the change handles a corner case in `colocated_cpu_devices()` API for using colocated Python on a PjRt-IFRT or IFRT Proxy on CPU (or any backend with CPU as the primary platform type). The API used to return a "cpu" backend's device even if the original backend device is on CPU, and had an inconsistent device selection between `colocated_cpu_devices()` and `Mesh`/`DeviceList` deserialization. Now these two APIs would prefer using the primary JAX backend's CPU devices.

PiperOrigin-RevId: 813948607