[Mosaic GPU] Improve checks on the WG strided transfers
Our previous code didn't actually verify that the vector loads/stores are safe,
so you could pass in a non-contiguous reference with a weird shape, but the right
number of elements and get nonsensical results. The current check is a bit too conservative,
but it's better to lean this way.

This also factors the address calculation into a common `transfer_strided` class method, following
the same pattern we use for tiled layouts.

PiperOrigin-RevId: 814163433