[Mosaic GPU] Conjure variables from `Divides` + Reg <-> SMEM `IsTransferable` constraints.
`IsTransferable` constraints are inserted whenever we copy data from one memory space to another. A subset of `IsTransferable` constraints will be about  Reg <-> SMEM copies. Whenever we see such a constraint, we can sometimes infer the SMEM transforms given the Register layout, provided it is an MMA layout. This change adds the logic that is needed to conjure assignments in those cases.

When conjuring variables, it is important to respect existing `Divides` constraints. For example, in the ragged dot case, we have an MMA layout that is accessed with a variety of different tiling constraints, where the major dimension of the tiles is powers of 2. In that kernel, we need to respect the `Divides` constraint and pick a tiling that works for all cases, which turns out to be a major tiling dimension of 1.

PiperOrigin-RevId: 814231869