Before sharding-in-types, the type of a Jacobian matrix came from concatenating the output and input types of the function being differentiated. For example, terms of shapes alone and setting aside pytrees, a function with signature `f32[a, b] -> f32[c, d]` has Jacobian matrix of shape `f32[c, d, a, b]`.
With sharding-in-types, we need a convention for how the Jacobian matrix's sharding is formed. We can't rely on concatenation in general, since the input and output of a function may both be sharded along the same mesh axis, so concatenating the types would be invalid. For example, for a function `f32[a@X] -> f32[b@X]` we can't form the Jacobian type `f32[b@X, a@X]`.

For jacrev we choose the convention that the Jacobian matrix type is formed by concatenating the output and input types while dropping the output sharding. For example, for a function `f32[a@X, b@Y] -> f32[c@X, d@Z]` we say the  jacrev Jacobian matrix has type `f32[c, d, a@X, b@Y]`. We need this convention because the VJP/pullback function has type `f32[c@X, d@Z] -> f32[a@X, b@Y]`, and with jacrev we compute the Jacobian by vmapping this pullback over the leading axes an identity matrix of type `f32[c, d, c@X, d@Z]` (either vmapping twice, or merging the first two axes and then vmapping once) to get a result `f32[c, d, a@X, b@Y]`.

For jacfwd, the convention differs for the same reason: for a function `f32[a@X, b@Y] -> f32[c@X, d@Z]` we say the jacfwd Jacobian matrix has type `f32[c@X, d@Z, a, b]`. That is, its type is formed by concatenating the output and input types while dropping the input sharding. That's because the JVP/pushforward function has type `f32[a@X, b@Y] -> f32[c@X, d@Z]`, and with jacfwd we compute the Jacobian by vmapping this pushforward over the trailing axes of an identity matrix of type `f32[a@X, b@Y, a, b]` to get a result `f32[c@X, d@Z, a, b]`.

If a user wanted a nontrivial sharding over the vmapped axes here, they can just make their own jacfwd or jacrev wrapper out of vmap and jvp/vjp, then control the sharding of the identity matrix.

PiperOrigin-RevId: 816533370