[Mosaic TPU] Improve the layouts used in the load+reshape fusion
Previously, the code would always end up loading Nx1x128 i32 chunks. This is obviously
inefficient as most of the vreg ends up as padding. The code also strided on the second minor,
which ended up being a complete no-op as only the first sublane was used...

The new code flattens the strided dimension into the dimension that ends up as the second minor
in the output, ensuring that we can actually perform the strided loads and produce data without
any padding.

PiperOrigin-RevId: 816616410