[Mosaic] Expand `arith.select` canonicalization
The function `canonicalize_select` only handles cases where the
condition is a scalar (i1 ? v1 : v2) by expanding the condition into a
vector using the operation `vector::BroadcastOp`. It does not handle
cases where the condition is a vector with element type bf16 and relies
on the function `canonicalize_bf16_vector` to do so.

Expand the function to handle bf16 canonicalization (and packed types
in general) as well. This would allow us to change the order of the
canonicalization rules to execute op-specific rules before general
canonicalization rules.

PiperOrigin-RevId: 816657256