Skip primitives without a defined roofline rule.
Previously, we'd raise a `NotImplementedError` if roofline encountered a primitive for which no roofline rule was registered. This was a useful tool during development so we could catch unimplemented primitives, but this approach is not very resilient. This change modifies the behavior to instead log a warning and skip the primitive, allowing roofline analysis to proceed for graphs containing such primitives.

PiperOrigin-RevId: 817240544