[XLA] Allow fine grain SparseCore offloading
Today any kind of SparseCore offloading is done as all or nothing, where this behaviour is controlled by a global flag. This is not always desirable, however the existing `compute_on` annotation to place ops on SparseCore was designed for embeddings in mind during TF days.

Introduce a new `compute_on` annotation `tpu_sparsecore_offload`, which will inform the compiler what ops should be offloaded. Note that we only support a handfull of ops for offloading and this method requires the offload annotation to survive the entire compilation pipeline, as offloading happens as late as possible. In case the annotation survives and the op (or whatever is left of it) was not offloaded, a compilation error will be returned.

PiperOrigin-RevId: 818694149