[Pallas:MGPU] Try to collapse dimension pairs in transfer_strided
Previously, our flattening scheme was all-or-nothing, where we were either
able to flatten all dimensions, or we would give up on it entirely. It's not
very difficult to make it finer-grained, however, and it lets us generalize
the method to be able to handle trailing dimensions that are not divisible
by the vector length (including singleton dimensions).

PiperOrigin-RevId: 819161012