[Mosaic GPU] Compress the representation of `Divides` constraints.
We don't need to carry information about dynamic variables if we resolve their
`gcd` at the creation site---which can be done since a multiple of every
possible tiling is always present (the axis of the given tiled dimension).

Some tests became obsolete as a result, and others could be simplified. This
allows to get rid of an MLIR dependency in `gpu_equations_test.py`, and also of
`nested_tuple`.

PiperOrigin-RevId: 821888942