Allow slice_sizes to gather that would exceed the operand dimension if the indices shape is empty.
This seems reasonable: it does not matter how large a slice is if you never actually collect any.

This shows up in cases like the following:

```
In [14]: s = jnp.zeros((0,), jnp.complex64)
In [15]: jax.jvp(partial(lax.sort, dimension=-1), [s], [s])

TypeError: Slice size at index 0 in gather op is out of range, must be within [0, 0 + 1), got 1.
```

Sort's JVP rule uses a gather to reorder the tangents in the same order as the primals. If there are no primals, there's nothing to gather, and it seems cleaner to allow gather to permit this case than to add special casing for the n=0 case.

PiperOrigin-RevId: 822819278