Implement Schur decomposition for Hessenberg matrices and return eigenvalues
Currently, only the call to LAPACK to calculate the Schur decomposition of a
general matrix has been implemented. This commit extends the
`jax.lax.linalg.schur` function to support Hessenberg matrix directly as
input and call the specific LAPACK code then.

Additionally, the function got the feature to return the eigenvalues from the
Schur form.