[Mosaic] Explain TPUv4 hotfix comment
cl/818758272 contained a hotfix that changed TPUv4's minimum supported bandwidth for the select operation.
At the time of its submission, the root cause was not clear, and a TODO comment was added to debug the failure further.

Now that the root cause related to hardware limitations around Vmsk retiling is understood,
this CL replaces the TODO comment with a more permanent comment detailing these limitations.

PiperOrigin-RevId: 824483978