Implement Krylov-Schur method for jax.scipy.sparse.linalg.eigs
In this commit the Krylov-Schur method as technically described in
https://slepc.upv.es/release/_downloads/5229480744b7c2533563dee75c16dfde/str7.pdf
is implemented as iterative solver for eigenvalue problems of general matrices.

The implementation mimics the scipy interface of scipy.sparse.linalg.eigs but is
not feature complete yet. Mainly, the implementation for generalized eigenvalue
problems is missing yet. Additionally, the implementation could be adapted for
hermitian/symmetric matrices which is a TODO for further work.