[Mosaic GPU] Fix op traversal during layout inference to be truly pre-order.
The fix also required to extend `consumer_operands` to also work with operands as input. This is the case e.g. for `ForOp` and `WhileOp` where operands that are mapped to block arguments effectively have consumers. Note I am just bringing this function inline with the existing logic of `producer_result` which already covered this special case.

PiperOrigin-RevId: 824505520