[Mosaic GPU] Refactor `Distinct` constraint in equation system.
The `Distinct` constraint is redefined to state that an expression is not an instance of a specific layout type (e.g., `WGSplatFragLayout`).

Requiring an instance of a specific layout is too restrictive. In a followup change, we would like to reject **all possible** strided and splat layouts in an equation system derivation rule.

PiperOrigin-RevId: 825110580