Add vary_unreduced_cast and reduced_vary_cast functions. `vary_unreduced_cast` is a `Varying -> Unreduced` cast and `reduced_vary_cast` is a `Reduced -> Varying` cast and **are transposes of each other**. They are no-ops at lowering time but during tracing time, they change the type of the output.
Note that these APIs are not yet exposed publicly. They are internal only and used in shard_map.py for now.

**Why?**

Consider this example:

```
mesh = jax.make_mesh((2, 2), ('x', 'y'))

@jax.shard_map(mesh=mesh, in_specs=(P('x', 'y'), P('y', None)),
               out_specs=P('x', None))
def f(x, y):
  return jnp.dot(x, y)

f(arr1, arr2)
```

The above code is incorrect and will error out because the out_specs is a lie! Since after the `dot`, we didn't reduce the output, the return value of the shard_map is in a unreduced state.

So if we need to make the above code work, we need to do `out_specs=P('x', None, unreduced={'y'})`. Let's write that down and see if it works:

```
@jax.shard_map(mesh=mesh, in_specs=(P('x', 'y'), P('y', None)),
               out_specs=P('x', None, unreduced={'y'}))
def f(x: f32[1, 2]{x,y}, y: f32[2, 2]{y}):
  z: f32[1, 2](x,y} = jnp.dot(x, y)
  return z

f(arr1, arr2)
```

The above code should work but without this change it won't. As you can see `z` is varying along both `x` and `y`, but out_specs is varying along `x` and unreduced along `y`.

Hence we need a cast to go from Varying -> Unreduced which is where `vary_unreduced_cast` API comes in. `reduced_vary_cast` is the transpose of `vary_unreduced_cast`.

`shard_map` in the above code will insert an implicit `vary_unreduced_cast` when it detects a varying->unreduced pattern like the above, via the `_implicit_unreduced_on_output` function.
This is similar to how we implicitly insert a pvary in similar cases via the `_implicit_pvary_on_output` function.

**Can this help get rid of psum_scatter_p and all_gather_p?**

Yes! `psum_scatter` can be implemented as `vary_unreduced_cast + unreduced_psum_scatter` and `all_gather` can be implemented as `all_gather_reduced + reduce_vary_cast`. This is so that we can maintain backwards compatibility.

This replacement can be prototyped and done in a follow up.

PiperOrigin-RevId: 825272725