[Mosaic GPU][NFC] Add the `Argument` VariableType.
This refactoring makes the logic a bit more straight-forward. Instead of special casing `producer_result` for `For` and `While` loops, the needed connection can appear directly in the corresponding rules for these ops. This also avoids special casing `consumer_operands` for `For` and `While` that would otherwise be needed for completeness.

Note that I am purposefully not renaming anything in this CL, in order to make the review easier. I will make the renaming changes in a follow up.

PiperOrigin-RevId: 825574070